Struct rusoto_ce::CostExplorerClient

pub struct CostExplorerClient { /* fields omitted */ }

A client for the AWS Cost Explorer API.

Methods

impl CostExplorerClient

pub fn new(region: Region) -> CostExplorerClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> CostExplorerClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,