Trait rusoto_ce::CostExplorer

pub trait CostExplorer {
    fn get_cost_and_usage(
        &self, 
        input: GetCostAndUsageRequest
    ) -> RusotoFuture<GetCostAndUsageResponse, GetCostAndUsageError>;

    fn get_cost_forecast(
        &self, 
        input: GetCostForecastRequest
    ) -> RusotoFuture<GetCostForecastResponse, GetCostForecastError>;

    fn get_dimension_values(
        &self, 
        input: GetDimensionValuesRequest
    ) -> RusotoFuture<GetDimensionValuesResponse, GetDimensionValuesError>;

    fn get_reservation_coverage(
        &self, 
        input: GetReservationCoverageRequest
    ) -> RusotoFuture<GetReservationCoverageResponse, GetReservationCoverageError>;

    fn get_reservation_purchase_recommendation(
        &self, 
        input: GetReservationPurchaseRecommendationRequest
    ) -> RusotoFuture<GetReservationPurchaseRecommendationResponse, GetReservationPurchaseRecommendationError>;

    fn get_reservation_utilization(
        &self, 
        input: GetReservationUtilizationRequest
    ) -> RusotoFuture<GetReservationUtilizationResponse, GetReservationUtilizationError>;

    fn get_tags(
        &self, 
        input: GetTagsRequest
    ) -> RusotoFuture<GetTagsResponse, GetTagsError>;
}

Trait representing the capabilities of the AWS Cost Explorer API. AWS Cost Explorer clients implement this trait.

Required methods

fn get_cost_and_usage(
&self,
input: GetCostAndUsageRequest
) -> RusotoFuture<GetCostAndUsageResponse, GetCostAndUsageError>

Retrieves cost and usage metrics for your account. You can specify which cost and usage-related metric, such as BlendedCosts or UsageQuantity, that you want the request to return. You can also filter and group your data by various dimensions, such as SERVICE or AZ, in a specific time range. For a complete list of valid dimensions, see the GetDimensionValues operation. Master accounts in an organization in AWS Organizations have access to all member accounts.

fn get_cost_forecast(
&self,
input: GetCostForecastRequest
) -> RusotoFuture<GetCostForecastResponse, GetCostForecastError>

Retrieves a forecast for how much Amazon Web Services predicts that you will spend over the forecast time period that you select, based on your past costs.

fn get_dimension_values(
&self,
input: GetDimensionValuesRequest
) -> RusotoFuture<GetDimensionValuesResponse, GetDimensionValuesError>

Retrieves all available filter values for a specified filter over a period of time. You can search the dimension values for an arbitrary string.

fn get_reservation_coverage(
&self,
input: GetReservationCoverageRequest
) -> RusotoFuture<GetReservationCoverageResponse, GetReservationCoverageError>

Retrieves the reservation coverage for your account. This enables you to see how much of your Amazon Elastic Compute Cloud, Amazon ElastiCache, Amazon Relational Database Service, or Amazon Redshift usage is covered by a reservation. An organization's master account can see the coverage of the associated member accounts. For any time period, you can filter data about reservation usage by the following dimensions:

AZ

CACHE_ENGINE

DATABASE_ENGINE

DEPLOYMENT_OPTION

INSTANCE_TYPE

LINKED_ACCOUNT

OPERATING_SYSTEM

PLATFORM

REGION

SERVICE

TAG

TENANCY

To determine valid values for a dimension, use the GetDimensionValues operation.

fn get_reservation_purchase_recommendation(
&self,
input: GetReservationPurchaseRecommendationRequest
) -> RusotoFuture<GetReservationPurchaseRecommendationResponse, GetReservationPurchaseRecommendationError>

Gets recommendations for which reservations to purchase. These recommendations could help you reduce your costs. Reservations provide a discounted hourly rate (up to 75%) compared to On-Demand pricing.

AWS generates your recommendations by identifying your On-Demand usage during a specific time period and collecting your usage into categories that are eligible for a reservation. After AWS has these categories, it simulates every combination of reservations in each category of usage to identify the best number of each type of RI to purchase to maximize your estimated savings.

For example, AWS automatically aggregates your Amazon EC2 Linux, shared tenancy, and c4 family usage in the US West (Oregon) Region and recommends that you buy size-flexible regional reservations to apply to the c4 family usage. AWS recommends the smallest size instance in an instance family. This makes it easier to purchase a size-flexible RI. AWS also shows the equal number of normalized units so that you can purchase any instance size that you want. For this example, your RI recommendation would be for c4.large because that is the smallest size instance in the c4 instance family.

fn get_reservation_utilization(
&self,
input: GetReservationUtilizationRequest
) -> RusotoFuture<GetReservationUtilizationResponse, GetReservationUtilizationError>

Retrieves the reservation utilization for your account. Master accounts in an organization have access to member accounts. You can filter data by dimensions in a time period. You can use GetDimensionValues to determine the possible dimension values. Currently, you can group only by SUBSCRIPTION_ID.

fn get_tags(
&self,
input: GetTagsRequest
) -> RusotoFuture<GetTagsResponse, GetTagsError>

Queries for available tag keys and tag values for a specified period. You can search the tag values for an arbitrary string.

Loading content...