Crate rusoto_ce

The Cost Explorer API allows you to programmatically query your cost and usage data. You can query for aggregated data such as total monthly costs or total daily usage. You can also query for granular data, such as the number of daily write operations for Amazon DynamoDB database tables in your production environment.

Service Endpoint

The Cost Explorer API provides the following endpoint:

https://ce.us-east-1.amazonaws.com

For information about costs associated with the Cost Explorer API, see AWS Cost Management Pricing.

If you're using the service, you're probably looking for CostExplorerClient and CostExplorer.

Structs

CostExplorerClient
A client for the AWS Cost Explorer API.

Coverage
The amount of instance usage that a reservation covered.

CoverageByTime
Reservation coverage for a specified period, in hours.

CoverageHours
How long a running instance either used a reservation or was On-Demand.

DateInterval
The time period that you want the usage and costs for.

DimensionValues
The metadata that you can use to filter and group your results. You can use GetDimensionValues to find specific values.

DimensionValuesWithAttributes
The metadata of a specific type that you can use to filter and group your results. You can use GetDimensionValues to find specific values.

EC2InstanceDetails
Details about the EC2 instances that AWS recommends that you purchase.

EC2Specification
The EC2 hardware specifications that you want AWS to provide recommendations for.

ESInstanceDetails
Details about the ES instances that AWS recommends that you purchase.

ElastiCacheInstanceDetails
Details about the ElastiCache instances that AWS recommends that you purchase.

Expression

Use Expression to filter by cost or by usage. There are two patterns:

Simple dimension values - You can set the dimension name and values for the filters that you plan to use. For example, you can filter for INSTANCETYPE==m4.xlarge OR INSTANCETYPE==c4.large. The Expression for that looks like this:

{ "Dimensions": { "Key": "INSTANCETYPE", "Values": [ "m4.xlarge", “c4.large” ] } }

The list of dimension values are OR'd together to retrieve cost or usage data. You can create Expression and DimensionValues objects using either with methods or set methods in multiple lines.

Compound dimension values with logical operations - You can use multiple Expression types and the logical operators AND/OR/NOT to create a list of one or more Expression objects. This allows you to filter on more advanced options. For example, you can filter on ((INSTANCETYPE == m4.large OR INSTANCETYPE == m3.large) OR (TAG.Type == Type1)) AND (USAGETYPE != DataTransfer). The Expression for that looks like this:

{ "And": [ {"Or": [ {"Dimensions": { "Key": "INSTANCETYPE", "Values": [ "m4.x.large", "c4.large" ] }}, {"Tags": { "Key": "TagName", "Values": ["Value1"] } } ]}, {"Not": {"Dimensions": { "Key": "USAGETYPE", "Values": ["DataTransfer"] }}} ] }

Because each Expression can have only one operator, the service returns an error if more than one is specified. The following example shows an Expression object that creates an error.

{ "And": [ ... ], "DimensionValues": { "Dimension": "USAGE_TYPE", "Values": [ "DataTransfer" ] } }

GetCostAndUsageRequest

GetCostAndUsageResponse

GetDimensionValuesRequest

GetDimensionValuesResponse

GetReservationCoverageRequest
You can use the following request parameters to query for how much of your instance usage is covered by a reservation.

GetReservationCoverageResponse

GetReservationPurchaseRecommendationRequest

GetReservationPurchaseRecommendationResponse

GetReservationUtilizationRequest

GetReservationUtilizationResponse

GetTagsRequest

GetTagsResponse

Group
One level of grouped data within the results.

GroupDefinition
Represents a group when you specify a group by criteria, or in the response to a query with a specific grouping.

InstanceDetails
Details about the instances that AWS recommends that you purchase.

MetricValue
The aggregated value for a metric.

RDSInstanceDetails
Details about the RDS instances that AWS recommends that you purchase.

RedshiftInstanceDetails
Details about the Amazon Redshift instances that AWS recommends that you purchase.

ReservationAggregates
The aggregated numbers for your Reserved Instance (RI) usage.

ReservationCoverageGroup
A group of reservations that share a set of attributes.

ReservationPurchaseRecommendation
A specific reservation that AWS recommends for purchase.

ReservationPurchaseRecommendationDetail
Details about your recommended reservation purchase.

ReservationPurchaseRecommendationMetadata
Information about this specific recommendation, such as the time stamp for when AWS made a specific recommendation.

ReservationPurchaseRecommendationSummary
A summary about this recommendation, such as the currency code, the amount that AWS estimates that you could save, and the total amount of reservation to purchase.

ReservationUtilizationGroup
A group of Reserved Instances (RIs) that share a set of attributes.

ResultByTime
The result that is associated with a time period.

ServiceSpecification
Hardware specifications for the service that you want recommendations for.

TagValues
The values that are available for a tag.

UtilizationByTime
The amount of utilization, in hours.

Enums

GetCostAndUsageError
Errors returned by GetCostAndUsage

GetDimensionValuesError
Errors returned by GetDimensionValues

GetReservationCoverageError
Errors returned by GetReservationCoverage

GetReservationPurchaseRecommendationError
Errors returned by GetReservationPurchaseRecommendation

GetReservationUtilizationError
Errors returned by GetReservationUtilization

GetTagsError
Errors returned by GetTags

Traits

CostExplorer
Trait representing the capabilities of the AWS Cost Explorer API. AWS Cost Explorer clients implement this trait.